chore: args organization
